turn simulated censored times as integers to create ties

+++ lights/simulation.py
@@ -507,4 +507,4 @@ class SimuJointLongitudinalSurvival(Simulation):
 
         self.long_features = Y
 
-        return X, Y, T, delta
+        return X, Y, T.astype(int), delta
